refactor: 인터셉터와 리졸버의 빈 등록 분리
